feat: update minor ✅

+++ src/model/user-dao.go
@@ -2,7 +2,7 @@ package model
 
 type UserDao struct {
 	Name        string `json:"name"`
-	Age         int    `json:"age"`
+	Age         int64  `json:"age"`
 	PhoneNumber string `json:"phoneNumber"`
 	Address     string `json:"address"`
 }
